audioflinger: Handle playback thread states
For an audio track that does not carry offload output flag,
as long as it runs on an offload thread, the calling logic of
pause/flush/destroy should be consistent with other audio
tracks with offload flag.
Add one judgment condition that thread type is OFFLOAD for
flush/destroy methods of audio track.

Test: atest AudioTrackTest
- Functional test - use exoplayer application to play multichannel
  AAC content then switches to play Dolby format content, both content
  playback are normal.
- No regressions found in basic playback tests.

Change-Id: I9d9e41492824c9c43065dd723f9cb78d1a476523
Signed-off-by: minaripenguin <minaripenguin@users.noreply.github.com>
Signed-off-by: Dmitrii <bankersenator@gmail.com>